Fix parsing of regexp options like --override=bin|man
Previously the \A and \z anchoring for --ignore / --override / --defer
was being applied without wrapping the user-specified regexp in
parentheses, so if it was something like bin|man then the anchor would
only apply to one half of those two paths.  So add parentheses, and
also change from qr() to qr{} syntax to visually differentiate between
regexp boundaries and parentheses inside the regexp.