ignore based on package dirent name prior to dot- adjustment
If `--dotfiles` was enabled, files in the package such as
`dot-gitignore` would be translated to `.gitignore` and then ignored
by the default ignore list.  However any file named `dot-*` in a
package is obviously intended to be stowed as a dot file, so should
not be ignored.

To fix this, ignore based on the name in the package, not the
potentially translated name used for stowing.