reposition Stow's description for its main use case
It's clear from YouTube videos, blogs, social media etc. that the
dotfiles management use case is by far the most popular one for Stow
these days, and also the labelling "symlink farm manager" didn't
really resonate - probably because it's too focused on the
implementation rather than the utility.

So rewrite the docs to emphasise the narrative which already resonates
best, whilst preserving acknowledgements that Stow can still be used
for other purposes, and the history behind what it was originally
built for.